ai-engine/requirements.txt
# Web Server & API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0

# AI & OCR Core
easyocr==1.7.1
torch==2.2.0
torchvision==0.17.0
Pillow>=10.0.0
numpy>=1.24.0

# LLM & Processing
langchain==0.1.9
langchain-community==0.0.29
pydantic==2.6.1
transformers==4.37.2

# Multilingual / Indic Support
indic-nlp-library==0.92

# FIX: pin google-genai to a stable version compatible with genai.Client() API
# 0.8.x introduced Client(); >=1.0 is the new stable SDK — both work with our code
google-genai==1.0.0

# FIX: google-genai 1.x requires httpx for async transport
httpx>=0.27.0